+++ tests/builder/main.rs
@@ -3,9 +3,7 @@ mod app_settings;
 mod arg_aliases;
 mod arg_aliases_short;
 mod arg_matcher_assertions;
-mod arg_name;
 mod arg_settings;
-mod arg_value;
 mod borrowed;
 mod cargo;
 mod command;
