+++ tests/macros.rs
@@ -469,4 +469,60 @@ mod arg_impl {
         let actual = clap::arg_impl! { @char 'o' };
         assert_eq!(actual, expected);
     }
+
+    #[test]
+    // allow double quoted dashed arg name in square brackets (e.g ["some-arg"])
+    fn arg_name_dashed() {
+        let arg = clap::arg!(["some-arg"] "some arg");
+        assert_eq!(arg, clap::Arg::new("some-arg").help("some arg"));
+
+        let m = clap::Command::new("flag")
+            .arg(arg)
+            .try_get_matches_from(vec!["", "some-val"])
+            .unwrap();
+        assert!(m.is_present("some-arg"));
+        assert_eq!(m.value_of("some-arg").unwrap(), "some-val");
+    }
+
+    #[test]
+    // allow double quoted dashed arg value in triangle brackets (e.g <"some-val">)
+    // test in combination with short argument name (e.g. -v)
+    fn arg_value_dashed_with_short_arg() {
+        let arg = clap::arg!(-a <"some-val"> "some arg");
+        assert_eq!(
+            arg,
+            clap::Arg::new("some-val")
+                .short('a')
+                .long("arg")
+                .value_name("some-val")
+        );
+
+        let m = clap::Command::new("cmd")
+            .arg(arg)
+            .try_get_matches_from(vec!["", "-a", "val"])
+            .unwrap();
+        assert!(m.is_present("some-val"));
+        assert_eq!(m.value_of("some-val").unwrap(), "val");
+    }
+
+    #[test]
+    // allow double quoted dashed arg value in triangle brackets (e.g <"some-val">)
+    // test in combination with long argument name (e.g. --value)
+    fn arg_value_dashed_with_long_arg() {
+        let arg = clap::arg!(-a --arg <"some-val"> "some arg");
+        assert_eq!(
+            arg,
+            clap::Arg::new("arg")
+                .short('a')
+                .long("arg")
+                .value_name("some-val")
+        );
+
+        let m = clap::Command::new("cmd")
+            .arg(arg)
+            .try_get_matches_from(vec!["", "--arg", "some-val"])
+            .unwrap();
+        assert!(m.is_present("arg"));
+        assert_eq!(m.value_of("arg").unwrap(), "some-val");
+    }
 }
